requirements.txt
Flask==3.0.3
bcrypt==4.1.2
pyotp==2.9.0
qrcode==7.4.2
opencv-python==4.9.0.80
face_recognition==1.3.0
PyJWT==2.8.0
numpy==1.26.4
Pillow==10.3.0
